ai_service/requirements.txt
flask==2.2.3
flask-cors==3.0.10
gunicorn==20.1.0
python-dotenv==1.0.0
numpy==1.24.2
pandas==1.5.3
tensorflow==2.12.0
tensorflow-hub==0.13.0
transformers==4.27.4
pillow==9.5.0
xgboost==1.7.5
scikit-learn==1.2.2
ultralytics==8.0.20
requests==2.28.2
pydantic==1.10.7
opencv-python-headless==4.7.0.72 